requirements.txt
Pillow==4.2.1
plyer==1.3.1.dev0